refactor(tsdown-driver): split adapter into responsibility modules
